[Add] CurrentMonthView
