+++ Ahttiary/Views/Calendar/CalendarView.swift
@@ -14,20 +14,6 @@ struct CalendarView: View {
     }// body
 }// CalendarView
 
-struct CurrentMonthView: View {
-    
-    @EnvironmentObject var dateManager: DateViewModel
-    
-    var body: some View {
-        VStack {
-            Text("\(dateManager.date)")
-            Text("Hello")
-            Text("Hello")
-        }
-
-    }// body
-}// CurrentMonthView
-
 struct CalendarView_Previews: PreviewProvider {
     static var previews: some View {
         CalendarView()
